Add new parameters for context documents and ease queryTensor restriction (#1287)
Introduce improved handling of missing documents and embeddings in recommendation and search workflows. It adds new parameters `allow_missing_documents` and `allow_missing_embeddings` to allow the API to optionally skip errors when documents or their embeddings are missing.

In addition, we ease the restriction on the hybrid search parameter `queryTensor`. It can now be either None, {} or not provided.